+++ machine/src/micro_common/mod.rs
@@ -642,7 +642,7 @@ impl DeviceInterface for LightMachine {
         for cpu_index in 0..cpu_topo.max_cpus {
             if cpu_topo.get_mask(cpu_index as usize) == 1 {
                 let thread_id = cpus[cpu_index as usize].tid();
-                let cpu_instance = cpu_topo.get_topo_instance_for_qmp(cpu_index as usize);
+                let cpu_instance = cpu_topo.get_topo_instance_for_qmp(cpu_index);
                 let cpu_common = qmp_schema::CpuInfoCommon {
                     current: true,
                     qom_path: String::from("/machine/unattached/device[")
@@ -683,10 +683,7 @@ impl DeviceInterface for LightMachine {
 
         for cpu_index in 0..self.base.cpu_topo.max_cpus {
             if self.base.cpu_topo.get_mask(cpu_index as usize) == 0 {
-                let cpu_instance = self
-                    .base
-                    .cpu_topo
-                    .get_topo_instance_for_qmp(cpu_index as usize);
+                let cpu_instance = self.base.cpu_topo.get_topo_instance_for_qmp(cpu_index);
                 let hotpluggable_cpu = qmp_schema::HotpluggableCPU {
                     type_: cpu_type.clone(),
                     vcpus_count: 1,
@@ -695,10 +692,7 @@ impl DeviceInterface for LightMachine {
                 };
                 hotplug_vec.push(serde_json::to_value(hotpluggable_cpu).unwrap());
             } else {
-                let cpu_instance = self
-                    .base
-                    .cpu_topo
-                    .get_topo_instance_for_qmp(cpu_index as usize);
+                let cpu_instance = self.base.cpu_topo.get_topo_instance_for_qmp(cpu_index);
                 let hotpluggable_cpu = qmp_schema::HotpluggableCPU {
                     type_: cpu_type.clone(),
                     vcpus_count: 1,
